// src/app/models/config.ts
export class Config {
    authenticationURL: string;
    searchPatientURL: string;
    patientAllDocsURL: string;
    patientSearchURL:string;
    searchProviderURL:string;
    providerAllDocsURL: string;
    userAllDocsURL: string;
    pharmacyAllDocsURL: string;
    labAllDocsURL: string;
    googleMapsURL: string; 
	googleMapsURLAppendOrigin: string;
	googleMapsURLAppendDestination: string; 
	googleMapsURLAppendDepartureTime : string;
	googleMapsURLAppendTrafficModel: string;
	googleMapsURLAppendKey : string;
	chaincodeId: string;
	chaincodeURL : string;
}